Intergrate new vision into wrapper
